+++ src/Moneybird.Net/Abstractions/IMoneybirdClient.cs
@@ -31,6 +31,16 @@ namespace Moneybird.Net.Abstractions
         IExternalSalesInvoiceEndpoint ExternalSalesInvoice { get; }
         
         /// <summary>
+        /// The FinancialAccount Endpoint.
+        /// </summary>
+        IFinancialAccountEndpoint FinancialAccount { get; }
+        
+        /// <summary>
+        /// The FinancialAccount Endpoint.
+        /// </summary>
+        IFinancialStatementEndpoint FinancialStatement { get; }
+        
+        /// <summary>
         /// The LedgerAccount Endpoint.
         /// </summary>
         ILedgerAccountEndpoint LedgerAccount { get; }