+++ tests/Moneybird.Net.Tests/Endpoints/FinancialStatements/FinancialStatementTestBase.cs
@@ -1,0 +1,6 @@
+namespace Moneybird.Net.Tests.Endpoints.FinancialStatements;
+
+public class FinancialStatementTestBase : CommonTestBase
+{
+    protected static string FinancialStatementId = "395773914725221835";
+}